Error 1213: Deadlock found when trying to get lock; try res tarting transaction

+++ core/reset.go
@@ -6,21 +6,23 @@ import (
 )
 
 func (fs *KFS) Reset(ctx context.Context, branchName string) error {
-	err := fs.Close()
-	if err != nil {
-		return err
+	if fs.isSqlite {
+		err := fs.Close()
+		if err != nil {
+			return err
+		}
+		err = os.RemoveAll(fs.root)
+		if err != nil {
+			return err
+		}
+		// TODO: fix reset with mysql
+		kfs, _, err := NewWithSqlite(fs.root, fs.newStorage)
+		if err != nil {
+			return err
+		}
+		fs.Db = kfs.Db
+		fs.S = kfs.S
 	}
-	err = os.RemoveAll(fs.root)
-	if err != nil {
-		return err
-	}
-	// TODO: fix reset with mysql
-	kfs, _, err := NewWithSqlite(fs.root, fs.newStorage)
-	if err != nil {
-		return err
-	}
-	fs.Db = kfs.Db
-	fs.S = kfs.S
-	_, err = fs.Checkout(ctx, branchName)
+	_, err := fs.Checkout(ctx, branchName)
 	return err
 }
